requirements.txt
Flask==1.1.2
requests==2.24.0
python-dotenv==0.13.0
tweepy==3.8.0
